#25 금액을 decimal로 변경

+++ Samples/Sample.AspNetCore/Controllers/CheckoutController.cs
@@ -98,7 +98,7 @@ namespace Sample.AspNetCore.Controllers
         {
             var preparation = await paymentsApi.PrepareAsync(new PaymentPreparation
             {
-                Amount = (int)amount,
+                Amount = amount,
                 TransactionId = transactionId,
             });
             if (preparation == null)
